refactor vault handlers
- co-locate input schemas & queries with handlers
- introduce `getPropsWithUserSession` to reduce boilerplate
